requirements.txt
flask
gunicorn
numpy
joblib
scikit-image
scikit-learn
pillow
opencv-python-headless